Enable macro tooltips
Hovering over a macro instantiation will show a the macro's signature and
a hyperlink to it's definition.
Examples:
    "#define MACRO 1" will show a tooltip "#define MACRO"
    "#define MACRO() 1" will show a tooltip "#define MACRO()"
    "#define MACRO(x, y) (x + y)" will show a tooltip "#define MACRO(x, y)"

To do this, we first need to parse with PARSE_DETAILED_PROCESSING_RECORD so
libclang will give us macro cursors.

Then, we need to open the file the macro is contained in and parse the
macro definition to get its signature.

Macro doxygen comments will not be shown, but could be added to MacroParser
in the future.