Implement spelling filters
These filters are applied to all completions and info popups. Their main
use is to make types and function signatures more readable.

Each filter consists of a regex pattern and what it should be replaced
with.

Example settings:
"spelling_filters": [
  {
    "pattern": " >",
    "replace": ">"
  },
  {
    "pattern": ", *std::allocator(<(?>[^<>]|(?1))*>)",
    "replace": ""
  }
]

If the regex module is available it is used instead of the re module.
This enables features like subroutine calls to search for matching
brackets as in the second example pattern.