Allow to specify custom header-to-source lookups
When using a compilation DB (i.e. directly or via CMake), one can now
use templates to find a "related" source file for a header file. This is
done by using globbing expressions which may contain additional
placeholders. These expressions describe how to look for a related
source file relative to a given header file. The first match that is
found in the DB is then used as flags source. This reduced the number of
flags passed to clang for auto-completion which in turn might avoid some
compilation issues in more complex project setups.