Use fnmatch to find related source files
Switching to use `fnmatch` instead of `glob`. This basically is required
to allow easier testing (without "real" files on disk). This might also
have a slight advantage regarding performance, as there are no
additional accesses to the disk to find related files - instead, we
search in the already loaded compilation DB for a related file.