Improved new option naming
Changed the naming of new options to make clear that we refer to the
target compiler (i.e. the one used in the user's build system) and not
clang itself which is used for generating completions. Also improved the
guessing of languages from a compiler invocation: We now also use the
file name to be able to detect Objective-C sources.