Remove filtering of "bad" defines
Remove the filtering of bad defines from compiler built ins. Clang is
fully able to continue code completion even when some of these system
defines are explicitly set. Hence, we try to pass them in explicitly to
ensure the view used by clang is as identical to that of the target
compiler as possible.