Popups for methods with arguments have hyperlinks to argument types
Example method signature:
    Foo* method(Bar* barArg);

Previously, popups for method() would include (if possible) a hyperlink
to the definition of the return type (Foo), but not any arguments (Bar).
Now, arguments such as Bar* will also be hyperlinked to the definition
of the Bar type.

Similar for ObjC/ObjC++ methods, the method below will now link to Bar
in addition to Foo:
    -(Foo*) method:(Bar*) bar;