Show an inline "error" popup instead of dialog
With this change, the error dialog is converted to a simple popup. This
is less intrusive than the dialog, as it does not block the user from
interacting with other parts of the GUI - on the other side, this is
more visible than just writing a log entry.

This contributes to #417 (by removing the dialog) but does not solve the
root cause (i.e. latest when editing a file where ECC uses combined
flags from build commands and hence we get multiple `-std=XXX` flags,
the error will be shown to the user).