compilation_db: fix file name generation
The specification of compile_commands.json from
https://clang.llvm.org/docs/JSONCompilationDatabase.html

Tells that the file has to be either absolut or relative to directory,
however, using the database file as relative directory sounds like a
good fallback to me.

This fixes wrong flags for all files when compile_commands.json are
generated with bear using 'bear make all'.