Introduce a flag source parsing include flags from Makefiles
This new flags source extracts Makefile variables that are commonly
used to specify include directories and other flags passed to the
compiler. This is especially useful for autotools based projects,
but may work for plain Makefile projects as well. It enables
completion for such projects without further configuration.

To extract the Makefile variables, make is invoked for the Makefile
next under the to-complete file. Using make target injection allows
us to delegate the rather complex Makefile parsing and just get an
easy-to-parse list of compiler flags.

For now this new "Makefile" flag sources is explicitly not yet added
to the default configuration, so we can test it on a wider range of
projects before potentially enabling it by default.