Refactor EasyClangComplete to show popup info using only class vars
Instead of using instance variables and instance methods
to kick off a thread job to get popup info and display it, refactor
to using class variables and static methods.

This will allow code outside of the EasyClangComplete instance
(for example, a new command to show popup info at a key press)
to show popup info.

Changes:
- Make current_job_id a class var instead of an instance var
- Make EasyClangComplete.info_finished a staticmethod
- Refactor EasyClangComplete.on_hover to use a staticmethod helper