Make valid syntax configurable
Close issue #432. This enables setting custom syntax to enable the plugin for
it. The downside is that now whenever we open a new view we have to load the
user settings which can take some minor time, but in practive it seems nearly
inperceptable.

This commit removes default syntaxes: "C++ (Colorcoded)", "cuda-c++", "deviot"